renamed jsx to js, added proptypes to callouts
