+++ .templates/new-module/src/js/react.js
