+++ .templates/new-module/tests/react/.babelrc
