+++ packages/accordion/tests/react/.babelrc
@@ -1,0 +1,3 @@
+{
+	"presets": ["react", ["es2015", { "modules": false }]]
+}
