+++ packages/callout/tests/react/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { Callout, CalloutCalendar } from './react.jsx';
+import { Callout, CalloutCalendar } from './react.js';
 
 ReactDOM.render(
 	<div>
@@ -11,7 +11,9 @@ ReactDOM.render(
 			Lorem ipsum dolor sit amet, consectetur adipisicing elit. Totam cupiditate ratione iste blanditiis, asperiores, recusandae, sed natus voluptate.
 		</Callout>
 
+		<hr />
 		<h2>Calendar callout</h2>
+
 		<CalloutCalendar
 			description="description for second callout"
 			datetime="2017-01-01T00:00:00+00:00"
@@ -20,7 +22,9 @@ ReactDOM.render(
 			name="Talk to boss"
 		/>
 
+		<hr />
 		<h2>Calendar callout without optionals</h2>
+
 		<CalloutCalendar
 			description="description for second callout"
 			datetime="2017-01-01T00:00:00+00:00"
